+++ src/runtime/RecoveredGameServicesRuntime.h
@@ -70,6 +70,7 @@ bool RecoveredGameServices_ExplosionSmokeReady();
 bool RecoveredGameServices_ExplosionPieceReady();
 bool RecoveredGameServices_ExplosionTraceReady();
 bool RecoveredGameServices_VehicleAttributesReady();
+bool RecoveredGameServices_VehicleReferencesReady();
 bool RecoveredGameServices_TaxiAttributesReady();
 bool RecoveredGameServices_TaxiReferencesReady();
 bool RecoveredGameServices_BulletAttributesReady();